Extract inner classes to separate files for readability
