+++ Assets/_Application/Layers/LayerTypes/GeoJsonLayers/GeoJSONPointLayer.MaterialApplicator.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 45d81ad478144187b80352b3204f921d
+timeCreated: 1744017884
