+++ Assets/_Application/Layers/LayerTypes/GeoJsonLayers/GeoJSONPolygonLayer.MaterialApplicator.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: dcc8115808244dbea2c718dd8e09d2f5
+timeCreated: 1744017573
