+++ Assets/_Application/Layers/MaterialApplicator.cs
@@ -1,5 +1,4 @@
-﻿using UnityEngine;
-using Object = UnityEngine.Object;
+﻿using Object = UnityEngine.Object;
 
 namespace Netherlands3D.Twin.Layers
 {
